Cleanup PHP doc comments
Return values
Type declarations etc.
